+++ steamclog/src/main/java/com/steamclock/steamclog/LogLevelPreset.kt
@@ -39,16 +39,16 @@ sealed class LogLevelPreset {
         get() = when(this) {
             is Firehose -> LogLevel.None
             is Develop -> LogLevel.None
-            is ReleaseAdvanced -> LogLevel.Info
-            is Release -> LogLevel.Warn
+            is ReleaseAdvanced -> LogLevel.Verbose
+            is Release -> LogLevel.Info
         }
 
     val sentry: LogLevel
         get() = when(this) {
             is Firehose -> LogLevel.None
             is Develop -> LogLevel.None
-            is ReleaseAdvanced -> LogLevel.Info
-            is Release -> LogLevel.Warn
+            is ReleaseAdvanced -> LogLevel.Verbose
+            is Release -> LogLevel.Info
         }
 
     val file: LogLevel